stock/requirements.txt
Flask[async]==3.0.2
redis==5.0.3
gunicorn==21.2.0
msgspec==0.18.6
psycopg2-binary
kafka-python==2.0.5
SQLAlchemy==2.0.38
flask-sqlalchemy==3.1.1